#68 - Moved the dice icon element and fixed the rich text element name
